+++ src/pages/summer/Login/Login.scss
@@ -21,31 +21,6 @@
           @include display-flex;
           font-size: 43px;
         }
-        form {
-          @include display-flex(unset, unset);
-          flex-direction: column;
-          margin-top: 45px;
-
-          input {
-            margin-bottom: 10px;
-            padding: 9px 0px 9px 8px;
-            border: 1px solid $color-lightgray;
-            border-radius: 2px;
-
-            input:focus {
-              border: 1px solid $color-lightgray;
-            }
-            input::placeholder {
-              font-size: 12px;
-            }
-          }
-          button {
-            margin: 5px 0 60px;
-            padding: 5px 0;
-            border-radius: 4px;
-            @include bold-text(white, 14px);
-          }
-        }
 
         .login-facebook {
           position: relative;
@@ -114,6 +89,7 @@
           }
           button {
             @include bold-text($color-active, 14px);
+            cursor: pointer;
           }
         }
       }
