+++ src/pages/summer/Login/LoginInput/LoginInput.scss
@@ -1,0 +1,27 @@
+@import '../../../../styles/common.scss';
+
+.login-container form {
+  @include display-flex(unset, unset);
+  flex-direction: column;
+  margin-top: 45px;
+
+  input {
+    margin-bottom: 10px;
+    padding: 9px 0px 9px 8px;
+    border: 1px solid $color-lightgray;
+    border-radius: 2px;
+
+    input:focus {
+      border: 1px solid $color-lightgray;
+    }
+    input::placeholder {
+      font-size: 12px;
+    }
+  }
+  button {
+    margin: 5px 0 60px;
+    padding: 5px 0;
+    border-radius: 4px;
+    @include bold-text(white, 14px);
+  }
+}
